opt: idr ring lock-free
